requirements.txt
torch
torchvision
opencv-python
numpy
ipython
matplotlib
pillow
pycocotools
datumaro
